requirements.txt
pyobjc-framework-Cocoa
pyqt5
pyyaml
slackclient
